+++ eosfactory/eosf.py
@@ -36,7 +36,7 @@ create_wallet = wallet.create_wallet
 get_wallet = wallet.get_wallet
 
 Account = account.Account
-MasterAccount = account.MasterAccount
+Account = account.Account
 create_account = account.create_account
 new_account = account.new_account
 create_master_account = account.create_master_account