+++ tests/tic_tac_toe.py
@@ -15,7 +15,7 @@ INITIAL_STAKE_NET = 3
 INITIAL_STAKE_CPU = 3
 
 # Actors of the test:
-MASTER = MasterAccount()
+MASTER = Account()
 HOST = Account()
 ALICE = None
 CAROL = None